pyproject.toml
[project]
name = "nonebot_plugin_dujie"
version = "0.1.0"
description = "一个基于NoneBot2的群聊修仙游戏插件"
authors = [
    {name = "biupiaa"}
]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "tortoise-orm>=0.19.0",
    "python-dotenv>=0.19.0",
    "aiosqlite>=0.17.0",
    "typing-extensions>=4.0.0",
    "pydantic>=1.8.0"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/ttlvip/nonebot_plugin_dujie"
Repository = "https://github.com/ttlvip/nonebot_plugin_dujie.git"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
test = [
    "pytest>=6.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0"
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = "E203" 